Some pedantic C++ fixes

+++ include/rive/factory.hpp
@@ -30,14 +30,14 @@ namespace rive {
                                                      float ex, float ey,
                                                      const ColorInt colors[],    // [count]
                                                      const float stops[],        // [count]
-                                                     int count,
+                                                     size_t count,
                                                      RenderTileMode,
                                                      const Mat2D* localMatrix = nullptr) = 0;
 
         virtual rcp<RenderShader> makeRadialGradient(float cx, float cy, float radius,
                                                      const ColorInt colors[],    // [count]
                                                      const float stops[],        // [count]
-                                                     int count,
+                                                     size_t count,
                                                      RenderTileMode,
                                                      const Mat2D* localMatrix = nullptr) = 0;
